delete createlist component, now included in home component
