+++ src/components/Home.js
@@ -1,11 +1,20 @@
 import React from 'react';
-import CreateList from './CreateList';
+import { NavLink } from 'react-router-dom';
+import getToken from '../lib/tokens';
 
 const Home = () => {
+  const newList = () => {
+    const token = getToken();
+    localStorage.setItem('token', token);
+  };
   return (
     <div>
       <h1>Welcome to Smart Shopping App</h1>
-      <CreateList />
+      <NavLink to="/list">
+        <button type="submit" onClick={newList}>
+          Create List
+        </button>
+      </NavLink>
     </div>
   );
 };
